Printify updates seem to work TODO: check for shipped event updates to add in shipping info? added vendorPO to udpatedata DTO added service to add that info. may need other data?

+++ MWW_MagicAPI/Data/Models/Magic/DAP_Partners.cs
@@ -7,7 +7,7 @@ namespace MWW_Api.Models.Magic;
 public class DapPartner
 {
     [Key]
-    public int UNIQ_ID { get; set; }
+    public decimal UNIQ_ID { get; set; }
     public string? PO { get; set; }
     public string? BILL_NAME { get; set; }
     public string? BILL_ADDRESS1 { get; set; }
